add per-arm belt extensions/stops This number is the length past where retraction stops that the belt actually extends.
This can be from putting a stop on the belt so it doesn't retract
all the way (useful so that you can do retraction without having
to disconnect from anchors) or an added length added to the belt
to be able to reach additional distances

In both cases, you figure out what value to put in by retracting
with and without the modification

if you are adding a clamp, retract without the modification,
extend, add the clamp and then retract again. The distance reported
when the belt is fully retracted is the value to enter for that arm.

If you are adding an extension, clamp the router in place, hook
the belt to an anchor (further away than your extension), retract
until it stops, extend, hook up your extension, retract again. The
absolute value of the distance reported is what you enter for that
arm.